Preliminary PVRTC alpha support.
This fixes #17.

However, it's a preliminary pass that only really operates on the image
included in #16. There are likely better ways to optimize the alpha
used in the images, but I'm not sure what that is. Likely it's better
to use separate alpha dilation than the intensity dilation algorithm
currently used.